doc_analyser/requirements.txt
streamlit
PyPDF2==3.0.1
python-docx==0.8.11
google-cloud-aiplatform
google-cloud-storage
sendgrid
google-cloud-secret-manager
tiktoken==0.4.0
transformers==4.33.1
langchain==0.1.5
langchain-community
langchain-google-vertexai
streamlit-option-menu
google-auth-oauthlib
google-api-python-client
google-cloud-firestore
openai